refactor:Numbers 인스턴스가 보다 캡슐화 될 수 있도록 로직 이동
