+++ src/main/java/step1/number/Numbers.java
@@ -1,17 +1,30 @@
 package step1.number;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 public final class Numbers {
 
     private final List<Number> numbers;
+
+    public Numbers(String[] data) {
+        this(convertStringArrayToNumberList(data));
+    }
 
     public Numbers(List<Number> numbers) {
         if (isNull(numbers)) {
             throw new IllegalArgumentException("null 값을 입력했습니다.");
         }
         this.numbers = numbers;
+    }
+
+    private static final List<Number> convertStringArrayToNumberList(String[] data) {
+        return Arrays.stream(data)
+                .mapToInt(Integer::parseInt)
+                .mapToObj(Number::new)
+                .collect(Collectors.toList());
     }
 
     private final boolean isNull(List<Number> numbers) {
